requirements.txt
# requirements.txt
yfinance
pandas
openpyxl
python-dotenv
requests
openai
pyyaml